Support github enterprise server for issue management
